recipe-samples/images/*: Add multiconfig support
Is becoming common to have 64 bits kernel with 32 bits userspace
in certain solutions.

A the new rpb-multiconfig.bbclass is used to build a 64 bits kernel
and copy into a 32 bits image deployemnt handling module installation
in rootfs.

A couple of variables needs to be set in order to enable this
functionality (in our dragonboard410c-32):

RPB_MULTICONFIG = "1"
RPB_MULTICONFIG_MACHINE_KERNEL = "dragonboard410c"

Signed-off-by: Aníbal Limón <anibal.limon@linaro.org>